// src/interfaces/construction.ts
export interface Construction {
  id: number;
  name: string;
  code: string;
  results: [];
  construction_code: string;
  zipcode: string;
  prefecture: string;
  municipality: string;
  address1: string;
  address2: string;
  structureRc: string;
  structureS: string;
  structureSrc: string;
  structureWooden: string;
  groundFloor: number;
  basementFloor: number;
  siteArea: number;
  siteAreaInTsubo: number;
  buildingArea: number;
  buildingAreaInTsubo: number;
  totalFloorArea: number;
  totalFloorAreaInTsubo: number;
  commodity: string;
  construction_details: string;
  eastWestDivision: string;
  newConstruction: string;
  scheduledStartDate: string;
  scheduledEndDate: string;
  estimatedDate: string;
  orderDate: string;
  contractDate: string;
  updatedDate: string;
  completionDate: string;
  assumedArtificial: string;
  dr0: string;
  dr1: string;
  dr2: string;
  confirmCompletionDate: string;
  fireinspectionDate: string;
  brand: Brand;
  supplier: Supplier;
  company: ConstructionInfoType;
  costManageMenu: CostManageMenuType[];
}

export interface ConstructionDatail {
  id: number;
  construction_code: string;
  construction_name: string;
  construction_content: string;
  product: string;
  east_west_division: number;
  main_brand_id: number;
  sub_brand1_id: number;
  sub_brand2_id: number;
  site_area: number;
  building_area: number;
  total_floor_area: number;
  structure_rc: boolean;
  structure_s: boolean;
  structure_src: boolean;
  structure_wooden: boolean;
  ground_floor: number;
  underground_floor: number;
  zipcode: string;
  prefecture: string;
  municipality: string;
  address_1: string;
  address_2: string;
  phone_no: string;
  customer_id: number;
  customer_department_name: string;
  contractor_name: string;
  contract_email_address: string;
  order_amount: number;
  estimated_amount: number;
  consumption_tax: number;
  estimated_assumption_cost: number;
  order_assumption_cost: number;
  working_budget: number;
  assumption_artificial: number;
  cost_real_time: number;
  assessment_amount: number;
  remaining_budget: number;
  gross_profit_amount_1: number;
  gross_profit_rate_1: number;
  gross_profit_amount_2: number;
  gross_profit_rate_2: number;
  status: number;
  estimated_date: string;
  order_date: string;
  contract_date: string;
  schedule_construction_start_date: string;
  schedule_construction_end_date: string;
  construction_started_date: string;
  confirmation_application: string;
  fire_inspection_date: string;
  dr0: string;
  dr1: string;
  dr2: string;
  work_place_score: number;
  user_review: string;
  construction_completion_date: string;
  old_construction_code: string;
  new_repair_division: number;
  sales_amount: number;
  additional_increase_decrease: number;
  existing_order_amount: number;
  unordered_budget: number;
  expenses: number;
  total_expenses: number;
  assessment_turnover: number;
  assumption_gross_profit: number;
  company_name: string;
  updated_at: Date;
  customer: {
    company_name: string;
  };
  main_brand: {
    brand_name: string;
  };
  sub_brand1: {
    brand_name: string;
  };
  sub_brand2: {
    brand_name: string;
  };
  overview: {
    id: number;
  };
}
// ...
export interface ConstructionSummaryDetail {
  id: number;
  construction_code: string;
  construction_name: string;
  construction_content: string;
  product: string;
  east_west_division: number;
  main_brand_id: number;
  sub_brand1_id: number;
  sub_brand2_id: number;
  site_area: number;
  building_area: number;
  total_floor_area: number;
  structure_rc: boolean;
  structure_s: boolean;
  structure_src: boolean;
  structure_wooden: boolean;
  ground_floor: number;
  underground_floor: number;
  zipcode: string;
  prefecture: string;
  municipality: string;
  address_1: string;
  address_2: string;
  phone_no: string;
  customer_id: number;
  customer_department_name: string;
  contractor_name: string;
  contract_email_address: string;
  order_amount: number;
  estimated_amount: number;
  consumption_tax: number;
  estimated_assumption_cost: number;
  order_assumption_cost: number;
  working_budget: number;
  assumption_artificial: number;
  cost_real_time: number;
  assessment_amount: number;
  remaining_budget: number;
  gross_profit_amount_1: number;
  gross_profit_rate_1: number;
  gross_profit_amount_2: number;
  gross_profit_rate_2: number;
  status: number;
  estimated_date: string;
  order_date: string;
  contract_date: string;
  schedule_construction_start_date: string;
  schedule_construction_end_date: string;
  construction_started_date: string;
  confirmation_application: string;
  fire_inspection_date: string;
  dr0: string;
  dr1: string;
  dr2: string;
  work_place_score: number;
  user_review: string;
  construction_completion_date: string;
  old_construction_code: string;
  new_repair_division: number;
  sales_amount: number;
  additional_increase_decrease: number;
  existing_order_amount: number;
  unordered_budget: number;
  expenses: number;
  total_expenses: number;
  assessment_turnover: number;
  assumption_gross_profit: number;
  company_name: string;
  updated_at: Date;
  customer: {
    id: number;
    customer_code: string;
    collection_code: string;
    corporation_no: string;
    company_name: string;
    company_name_kana: string;
    branch_department_name: string;
    branch_department_kana: string;
    zipcode: string;
    prefecture: string;
    municipality: string;
    address_1: string;
    address_2: string;
    phone_no: string;
    email: string;
    invoice_no: string;
    fax: string;
    url: string;
    representative_name: string;
    representative_furigana: string;
    use_suspension_flag: string;
    suspension_reason: string;
  };
  construction_info: {
    rule: string;
    map: string;
    movie: string;
  };
  overview: {
    id: number;
  };
}