// src/types/api/construction.ts
export type TypeConstructionList = {
  id: number;
  construction_code: string;
  construction_name: string;
  municipality: string;
  status: number;
  customer: {
    id: number;
    company_name: string;
  };
  main_brand: {
    id: number;
    brand_id: string;
    brand_icon: string;
  };
  sub_brand1: {
    id: number;
    brand_id: string;
  };
  sub_brand2: {
    id: number;
    brand_id: string;
  };
}[];